// packages/core/src/QueryBuilder/types.d.ts
export type QueryRuleValue =
  | string
  | number
  | boolean
  | NumericRange
  | DateTimeStrings
  | DateTimeRange;

export interface QueryRule {
  id?: number | string;
  attribute?: string;
  operator?: string;
  value?: QueryRuleValue;
}

export interface Query {
  id?: number;
  combinator: string;
  rules: Array<QueryRule>;
}

export interface QueryCombinator {
  operand: string;
  label: string;
}

export interface QueryOperator {
  operator: string;
  label: string;
  combinators: string[];
}

interface DialogLabels {
  dialogTitle: string;
  dialogMessage: string;
  dialogConfirm: string;
  dialogCancel: string;
  dialogCloseTooltip: string;
}

interface ResetQueryAction {
  type: "reset-query";
}

interface ResetGroupAction {
  type: "reset-group";
  id: number;
}

interface AddRemoveAction {
  type: "add-rule" | "add-group" | "remove-node";
  id: number;
}

interface SetCombinatorAction {
  type: "set-combinator";
  id: number;

  combinator: string;
}

interface SetAttributeAction {
  type: "set-attribute";
  id: number;

  attribute: string | null;
  operator?: string | null;
  value?: QueryRuleValue | null;
}

interface SetOperatorAction {
  type: "set-operator";
  id: number;

  operator: string | null;
  value?: QueryRuleValue | null;
}

interface SetValueAction {
  type: "set-value";
  id: number;

  value: QueryRuleValue | null;
}

export type QueryAction =
  | ResetQueryAction
  | ResetGroupAction
  | AddRemoveAction
  | SetCombinatorAction
  | SetAttributeAction
  | SetOperatorAction
  | SetValueAction;